added command for using GMM as opposed to KMeans
