+++ EnrichSeq.py
@@ -51,6 +51,7 @@ def parseArgs(argv=None) -> argparse.Namespace:
     enrich_parser.add_argument("-kdb", "--kracken_db", help="the path to the kraken database", required=False)
     enrich_parser.add_argument("-gdb", "--genome_db", help="the path to the phage genome directory", required=False)
     enrich_parser.add_argument("-v", "--verbose", action="store_true", help="prints output figures and debug info", required=False)
+    enrich_parser.add_argument("-c", "--use_gmm", action="store_false", help="use Gaussian Mixture Model in Merge Overlap", required=False)
     enrich_parser.add_argument("-t", "--threads", help="number of threads to use [Default 4]", required=False)
     return parser.parse_args(argv)
 
@@ -91,6 +92,8 @@ def run_enrichseq(primary_args):
         CMD_list += ["--genomedir", f"{CURR_PATH}/database/ref_genomes/"]
     if primary_args.verbose:
         CMD_list += ["--verbose", "True"]
+    if primary_args.use_gmm:
+        CMD_list += ["--use_gmm"]
     CMD_list += ["--workdir", primary_args.output]
     out = utils.subproc_call(CMD_list)
     return out
