+++ src/nextflow/nextflow.config
@@ -2,6 +2,7 @@ params {
 	help = false
 	verbose = false
 	read = "single"
+	use_gmm = false
 	fasta = ""
 	workdir = ""
 	genomedir = "database/ref_genomes/"
